Social is almost over. User profile added. And some changes.

+++ app/src/main/java/com/example/mygoaldiary/Helpers/UserTasksHelpers/DeleteTask.kt
@@ -6,9 +6,8 @@ import android.content.Context
 import android.view.View
 import android.widget.CheckBox
 import com.example.mygoaldiary.Creators.DeleteAlertDialog
-import com.example.mygoaldiary.Creators.ShowAlert
-import com.example.mygoaldiary.Details
-import com.example.mygoaldiary.Fragments.Fragments.HomeMenuFragments.UserProjects
+import com.example.mygoaldiary.Views.Details
+import com.example.mygoaldiary.Views.HomeMenuFragments.UserProjects.UserProjects
 import com.example.mygoaldiary.Helpers.InternetController
 import com.example.mygoaldiary.R
 
@@ -17,11 +16,19 @@ class DeleteTask (private val mContext:  Context, private val mActivity : Activi
     private lateinit var alertDialogHere: AlertDialog
 
     fun delete(deleteProjectView: View, taskUuid : String?, isHybrid : String){
+
         loadingDialog.startLoadingDialog()
         alertDialogHere = DeleteAlertDialog.alertDialog
         val isChecked = deleteProjectView.findViewById<CheckBox>(R.id.deleteInternetTooCheckBox).isChecked
 
         if (isChecked) {
+
+            if (currentUser == null){
+                loadingDialog.dismissDialog()
+                showAlert.errorAlert("Error", "You must be logged in to delete a task from the cloud.", true)
+                return
+            }
+
             if (isHybrid == "true") {
                 if (InternetController.internetControl(mActivity)) {
                     deleteFromFirebase(taskUuid)
@@ -57,13 +64,13 @@ class DeleteTask (private val mContext:  Context, private val mActivity : Activi
     }
 
     private fun deleteFromSQL(taskUuid: String?) {
-        val get = sqlManage.manager(mSql, "DELETE FROM '${Details.projectId}' WHERE  taskUuid = '$taskUuid'")
+        val get = sqlManage.manager(mSql, "DELETE FROM '${Details.projectUuid}' WHERE  taskUuid = '$taskUuid'")
         if (get){
             alertDialogHere.cancel()
             refreshAllViewsFromTasksLayout(mContext, mActivity)
         }else{
             alertDialogHere.cancel()
-            showAlert.errorAlert(getString(R.string.error), getString(R.string.deleteTaskFail), true)
+            showAlert.errorAlert(mContext.getString(R.string.error), mContext.getString(R.string.deleteTaskFail), true)
         }
         loadingDialog.dismissDialog()
     }
